[#76596764] Exports raw blender shadernodes
The future of blender rendering is shadernodes : a node system that allows the artist to manipulate different shaders and create the perfect rendering without a line of code.
This shadernode tree is mostly exportable into an osg file userdata.
It can then later be read and compiled into a real live shader.
Some nodes will not be supported (refraction, ambient occlusion, ...) because hard or even impossible in realtime.