[#84276452] Modifies visibity checking
The previous behaviour of the exporter was to skip all the children of
a non visible object. As visibility is not propagated within the hierarchy,
objects are visible even if their parent is not, so we have to take them
into account. Another reason to avoid discarding not visible objects is that
transformations have to be kept