[#54671526] Makes objects instances (by vertex) as real objects
When the DupliVerts tool is used in Blender on an object, its children are
duplicated according to each of its vertexes (i.e an instance of the children is
created and 'snapped' on each of the vertexes of the parent).
We now create a real object for each instance in order to retrieve these geometries.
The 'duplicates_make_real()' allows to do this.