[#D3D-1600] Doesn't export influence if no armature modifier
Seems like it was an issue to export influences for mesh children of armature
since the relation can be only for transform. The code was commented before,
and it has probably been wrongly re-enabled.
Blender always add an Armature modifier for any parenting including rigging,
so it seems to be a reliable way to consider an object as rigged.

If any issue with influcences not exported, it will need a better solution